Renamed AddConfig.py to add_config.py
